DEV-9588 API-15 : merge 7526 from trunk
git-svn-id: http://svn.terracotta.org/svn/ehcache@7527 b9324663-ca0f-0410-8574-be9b3887307d

+++ branches/ehcache-2.7.x/ehcache-core/src/main/java/net/sf/ehcache/config/AbstractCacheConfigurationListener.java
@@ -86,7 +86,7 @@ public abstract class AbstractCacheConfigurationListener implements CacheConfigu
      * {@inheritDoc}
      */
     @Override
-    public void maxEntriesInCacheChanged(int oldCapacity, int newCapacity) {
+    public void maxEntriesInCacheChanged(long oldCapacity, long newCapacity) {
         //
     }
