+++ branches/ehcache-2.7.x/ehcache-core/src/test/java/net/sf/ehcache/config/CacheConfigurationListenerTest.java
@@ -298,7 +298,7 @@ public class CacheConfigurationListenerTest extends AbstractCacheTest {
         /**
          * {@inheritDoc}
          */
-        public void maxEntriesInCacheChanged(final int oldValue, final int newValue) {
+        public void maxEntriesInCacheChanged(final long oldValue, final long newValue) {
             // no-op
         }
     }
